Created first version of the compact calendar
The first version of the compact calendar component is done. There are a lot of CSS styling update that need to be done for the compact version to work well.
Also added color coding to the event dots so a user will know what types of events are on what day at a glance